chore #13: move component to presentation layer
