+++ app/src/main/java/org/sopt/and/feature/component/RoundedButton.kt
@@ -1,10 +1,11 @@
-package org.sopt.and.component
+package org.sopt.and.feature.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -19,7 +20,7 @@ import org.sopt.and.ui.theme.Gray
 import org.sopt.and.ui.theme.White
 
 @Composable
-fun ExpandedButton(
+fun RoundedButton(
     text: String,
     onClick: () -> Unit,
     modifier: Modifier = Modifier,
@@ -32,6 +33,7 @@ fun ExpandedButton(
             .clickable { if (enabled) onClick() }
             .background(
                 color = if (enabled) Blue else Gray,
+                shape = RoundedCornerShape(50.dp)
             )
     ) {
         Text(
@@ -45,9 +47,9 @@ fun ExpandedButton(
 
 @Preview
 @Composable
-fun ExpandedButtonPreview() {
+fun RoundedButtonPreview() {
     ANDANDROIDTheme {
-        ExpandedButton(
+        RoundedButton(
             text = "",
             onClick = {}
         )
