+++ app/src/main/java/org/sopt/and/feature/home/HomeRoute.kt
@@ -33,8 +33,8 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
 import org.sopt.and.R
-import org.sopt.and.component.HomeContent
-import org.sopt.and.component.TopBar
+import org.sopt.and.feature.component.HomeContent
+import org.sopt.and.feature.component.TopBar
 import org.sopt.and.feature.home.model.ContentModel
 import org.sopt.and.ui.theme.ANDANDROIDTheme
 import org.sopt.and.ui.theme.Black
