+++ app/src/main/java/org/sopt/and/feature/signin/SignInRoute.kt
@@ -37,9 +37,9 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
 import org.sopt.and.R
-import org.sopt.and.component.RoundedButton
-import org.sopt.and.component.AuthTextField
-import org.sopt.and.component.TopBar
+import org.sopt.and.feature.component.RoundedButton
+import org.sopt.and.feature.component.AuthTextField
+import org.sopt.and.feature.component.TopBar
 import org.sopt.and.ui.theme.ANDANDROIDTheme
 import org.sopt.and.ui.theme.Black
 import org.sopt.and.ui.theme.LightGray
