+++ app/src/main/java/org/sopt/and/feature/signup/SignUpRoute.kt
@@ -31,9 +31,9 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
 import org.sopt.and.R
-import org.sopt.and.component.ExpandedButton
-import org.sopt.and.component.AuthTextField
-import org.sopt.and.component.TopBar
+import org.sopt.and.feature.component.ExpandedButton
+import org.sopt.and.feature.component.AuthTextField
+import org.sopt.and.feature.component.TopBar
 import org.sopt.and.showToast
 import org.sopt.and.ui.theme.ANDANDROIDTheme
 import org.sopt.and.ui.theme.Black
